Move dataTypes to individual files
