+++ src/app/datatypes/Project.ts
@@ -1,0 +1,8 @@
+export interface Project {
+   id: number;
+   title: string;
+   description: string;
+   goal: string;
+   authorID: number;
+   userIDs: number[];
+}
