+++ src/app/datatypes/Task.ts
@@ -1,0 +1,10 @@
+export interface Task {
+   id: number;
+   name: string;
+   description: string;
+   status: string;
+   authorID: number;
+   userIDs: number[];
+   projectID: number;
+   location: string;
+}
